Proposal: Creating the position of an APDA Webmaster and Enshrining the Tech Committee
Over the course of this year, APDA Board has come to the realization that both apda.online and NU tab have accrued a significant amount of “tech debt,” which, if addressed, would significantly improve the functioning of the league. Recognizing that addressing these issues would likely require a multi-year commitment, APDA board is proposing the creation of a “Webmaster” position to oversee these reforms and maintain the aforementioned softwares.